Fixes PR: Moved setting background image to activity-utils Change setting background image in theme buttons from ImageTheme to ImageBackground
