+++ src/components/theme-buttons.tsx
@@ -2,7 +2,8 @@
 // and is not intended to be displayed in a production release.
 // Add this component to enable buttons to toggle color theme.
 import React from "react";
-import { setThemeColors, setBackgroundImage } from "../utilities/theme-utils";
+import { setThemeColors } from "../utilities/theme-utils";
+import { setBackgroundImage } from "../utilities/activity-utils";
 
 import "./theme-buttons.scss";
 
@@ -22,7 +23,7 @@ export const ThemeButtons: React.FC = () => {
   const handleInteractionsTheme = () => {
     setThemeColors("#414546", "#2f70b0");
   };
-  const handleImageTheme = () => {
+  const handleImageBackground = () => {
     setBackgroundImage("https://upload.wikimedia.org/wikipedia/commons/thumb/0/0c/Hubble%27s_Wide_View_of_%27Mystic_Mountain%27_in_Infrared.jpg/678px-Hubble%27s_Wide_View_of_%27Mystic_Mountain%27_in_Infrared.jpg");
   };
   return (
@@ -32,7 +33,7 @@ export const ThemeButtons: React.FC = () => {
       <button className="button" onClick={handleCBioTheme} data-cy="theme-button-cbio">cbio</button>
       <button className="button" onClick={handleWatersTheme} data-cy="theme-button-waters">waters</button>
       <button className="button" onClick={handleInteractionsTheme} data-cy="theme-button-interactions">interactions</button>
-      <button className="button" onClick={handleImageTheme} data-cy="theme-button-image">image</button>
+      <button className="button" onClick={handleImageBackground} data-cy="theme-button-image">image</button>
     </div>
   );
 };
